MEDIUM: quic: requeue datagrams received on wrong socket
There is a small race condition when QUIC connection socket is
instantiated between the bind() and connect() system calls. This means
that the first datagram read on the sockets may belong to another
connection.

To detect this rare case, we compare the DCID for each QUIC datagram
read on the QUIC socket. If it does not match the connection CID, the
datagram is requeue using quic_receiver_buf to be able to handle it on
the correct thread.

This change is part of quic-conn owned socket implementation.
It may be backported to 2.7 after a period of observation.

(cherry picked from commit b2bd83972b7f61f674871b9a9722a6192a841322)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>