DOC: configuration: missing 'if' in tcp-request content example
An example given for tcp-request content rule with lua
was missing 'if' keyword. Using it "as is" makes haproxy unhappy.

The example was introduced with 579d83b05.
So it may be backported as far as 1.6, but it is a really minor typo.

(cherry picked from commit d49b559a15a8781ba37877812d4d3ddf2cea9eb4)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>