BUILD: ssl_sock: bind_conf uninitialized in ssl_sock_bind_verifycbk()
Even if this cannot happen, ensure <bind_conf> is initialized in this
function to please some compilers.

Takes the opportunity of this patch to replace an ABORT_NOW() by
a BUG_ON() because if the variable values they test are not initialized,
this is really because there is a bug.

Must be backported to 2.6.

(cherry picked from commit 48e46f98ccf97427995eb41c6f28cc38705bdd7e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>