MINOR: httpclient/lua: Don't set req_payload callback if body is empty
The HTTPclient callback req_payload callback is set when a request payload
must be streamed. In the lua, this callback is set when a body is passed as
argument in one of httpclient functions (head/get/post/put/delete). However,
there is no reason to set it if body string is empty.

This patch is related to the issue #1898. It may be backported as far as
2.5.

(cherry picked from commit 380ae9c3ff4919c9ea36d3d713427575c44fe4b8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>