BUG/MINOR: log: Preserve message facility when the log target is a ring buffer
When a ring is used as log target, the original facility, if any, must be
preserved. The default facility must only be used if there no facility was
found in the incoming log message.

This patch should fix the issue #1901. It must be backported as far as 2.4.

(cherry picked from commit cc640e851aa22e9ba7c4b5155d099457d1e3989f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>