REGTESTS: httpclient/lua: test the lua task timeout with the httpclient
Test the httpclient when the lua action timeout. The lua timeout is
reached before the httpclient is ended. This test that the httpclient
are correctly cleaned when destroying the hlua context.

Must be backported as far as 2.5.

(cherry picked from commit 4ed0a3a88384a3836cbf162025df7ca41cc4fa5e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>