BUG/MINOR: stick-table: Use server_id instead of std_t_sint in process_store_rules()
Te commit 8c3cdf0a99 ("BUG/MAJOR: stick-tables: do not try to index a server
name for applets") was not properly backported. On HAProxy 2.4, the server
id have to be store in the server_id data type.

It is a 2.4-specific commit. Thus there is no upstream ID. No backport
needed.