BUG/MAJOR: stick-table: don't process store-response rules for applets
The commit bc7c207f74 ("BUG/MAJOR: stick-tables: do not try to index a
server name for applets") tried to catch applets case when we tried to index
the server name. However, there is still an issue. The applets are
unconditionally casted to servers and this bug exists since a while. it's
just luck if it doesn't crash.

Now, when store rules are processed, we skip the rule if the stream's target
is not a server or, of course, if it is a server but the "non-stick" option
is set. However, we still take care to release the sticky session.

This patch must be backported to all stable versions.

(cherry picked from commit b976640fe178c7084f0db4d443874ebe4746dca8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 4c629b75a362817f57cdcc5f41bc79d20aac362a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 738b47e029c25f2cd7c77a75ea377eb37912b20f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>