BUG/MINOR: resolvers: Set port before IP address when processing SRV records
For a server subject to SRV resolution, when the server's address is set,
its dynamic cookie, if any, and its server key are computed. Both are based
on the ip/port pair. However, this happens before the server's port is
set. Thus the port is equal to 0 at this stage. It is a problem if several
servers share the same IP but with different ports because they will share
the same dynamic cookie and the same server key, disturbing this way the
connection persistency and the session stickiness.

This patch must be backported as far as 2.2.

(cherry picked from commit 2364b39984e4fd4aa6a88148520d49fe0620c034)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 24e3037b525eb8630e41edc7d69774ef5014464e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>