BUG/MINOR: http-htx: Fix error handling during parsing http replies
When an error is triggered during arguments parsing of an http reply (for
instance, from a "return" rule), while a log-format body was expected but
not evaluated yet, HAproxy crashes when the body log-format string is
released because it was not properly initialized.

The list used for the log-format string must be initialized earlier.

This patch should fix the issue #1925. It must be backported as far as 2.2.

(cherry picked from commit 5a3d9a77e20a66796357627b1dd807a1b22603a5)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>