MINOR: quic: define config option for socket per conn
Define global configuration option "tune.quic.socket-owner". This option
can be used to activate or not socket per QUIC connection mode. The
default value is "listener" which disable this feature. It can be
activated with the option "connection".

This change is part of quic-conn owned socket implementation.
It may be backported to 2.7 after a period of observation.

(cherry picked from commit 511ddd5785266c149dfa593582512239480e1688)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>